+++ omnifeed-unsplash/src/commonTest/kotlin/io/lackstudio/omnifeed/unsplash/domain/UnsplashUseCaseTest.kt
@@ -12,8 +12,8 @@ import io.lackstudio.omnifeed.core.network.KtorConfig
 import io.lackstudio.omnifeed.core.network.oauth.AccessTokenProvider
 import io.lackstudio.omnifeed.unsplash.data.remote.api.UnsplashApiService
 import io.lackstudio.omnifeed.unsplash.data.remote.api.UnsplashApiServiceImpl
-import io.lackstudio.omnifeed.unsplash.data.local.LocalUnsplashPhotoDataSource
-import io.lackstudio.omnifeed.unsplash.data.model.response.PhotoDetailResponse
+import io.lackstudio.omnifeed.unsplash.data.local.source.LocalUnsplashPhotoDataSource
+import io.lackstudio.omnifeed.unsplash.data.remote.model.response.PhotoDetailResponse
 import io.lackstudio.omnifeed.unsplash.data.remote.source.RemoteUnsplashDataSource
 import io.lackstudio.omnifeed.unsplash.data.remote.source.RemoteUnsplashDataSourceImpl
 import io.lackstudio.omnifeed.unsplash.data.repository.UnsplashRepositoryImpl
